fix: address source-file-lengths and vscode-art version CI failures
- Bump editors/vscode-art/package.json version 0.251.0 → 0.252.0
- Move module augmentation from nuxt/src/index.ts to options.ts (net-zero)
- Extract resolveMuseaSharedConfig to vite-musea/src/plugin/config.ts
- Collapse expanded import in vite/src/plugin/index.ts
- Extract NewArgs + run_new to crates/vize/src/commands/musea/new.rs
